begin with seekerbar

+++ 23Ballenger/engine/Fontrenderer.cpp
@@ -46,28 +46,28 @@ void Fontrenderer::setCamera(const Camera& camera) {
 	m_batchrenderer->setCamera(camera);
 }
 
-void Fontrenderer::drawText(const CharacterSet& characterSet, int posX, int posY, std::string text, Vector4f color) {
+void Fontrenderer::drawText(const CharacterSet& characterSet, float posX, float posY, std::string text, Vector4f color) {
 	glBindTexture(GL_TEXTURE_2D_ARRAY, characterSet.spriteSheet);
 	std::string::const_iterator c;
 	for (c = text.begin(); c != text.end(); c++) {
 
 		const Char& ch = characterSet.getCharacter(*c);
 
-		m_batchrenderer->addQuadAA(Vector4f(static_cast<float>(posX), static_cast<float>(posY), static_cast<float>(ch.size[0]), static_cast<float>(ch.size[1])), Vector4f(ch.textureOffset[0], ch.textureOffset[1], ch.textureSize[0], ch.textureSize[1]), color, characterSet.frame);
+		m_batchrenderer->addQuadAA(Vector4f(posX, posY, static_cast<float>(ch.size[0]), static_cast<float>(ch.size[1])), Vector4f(ch.textureOffset[0], ch.textureOffset[1], ch.textureSize[0], ch.textureSize[1]), color, characterSet.frame);
 		posX = posX + ch.advance;
 	}
 	m_batchrenderer->drawBuffer();
 	glBindTexture(GL_TEXTURE_2D_ARRAY, 0);
 }
 
-void Fontrenderer::addText(const CharacterSet& characterSet, int posX, int posY, std::string text, Vector4f color) {
+void Fontrenderer::addText(const CharacterSet& characterSet, float posX, float posY, std::string text, Vector4f color) {
 	std::string::const_iterator c;
 
 	for (c = text.begin(); c != text.end(); c++) {
 
 		const Char& ch = characterSet.getCharacter(*c);
 
-		m_batchrenderer->addQuadAA(Vector4f(static_cast<float>(posX), static_cast<float>(posY), static_cast<float>(ch.size[0]), static_cast<float>(ch.size[1])), Vector4f(ch.textureOffset[0], ch.textureOffset[1], ch.textureSize[0], ch.textureSize[1]), color, characterSet.frame);
+		m_batchrenderer->addQuadAA(Vector4f(posX, posY, static_cast<float>(ch.size[0]), static_cast<float>(ch.size[1])), Vector4f(ch.textureOffset[0], ch.textureOffset[1], ch.textureSize[0], ch.textureSize[1]), color, characterSet.frame);
 		posX = posX + ch.advance;
 	}
 }
